feat: introduce recon_operator package with autonmap compatibility shim
Move the Quart server into recon_operator.server, keep import/CLI paths via
autonmap aliasing, and expose auth/jobs/scheduler/api/config package surfaces
for the P2-01 modular layout (v1.9.0).
